third_party/blink/web_tests/html/selectmenu/selectmenu-typeahead-disabled-option-expected.txt

Typeahead does not select disabled selectmenu option

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS selectmenu.open is false
PASS selectmenu.value is "United States"
PASS successfullyParsed is true

TEST COMPLETE
